% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormmixcode.R
\name{dnormmix}
\alias{dnormmix}
\title{PDF for Mixture of Normal Distributions}
\usage{
dnormmix(x, mean1, sd1, mean2, sd2, mixprop)
}
\arguments{
\item{x}{A given quantile for the PDF.}

\item{mean1}{Mean for the first normal distribution.}

\item{sd1}{Standard deviation for the first normal distribution.}

\item{mean2}{Mean for the second normal distribution.}

\item{sd2}{Standard deviation for the second normal distribution.}

\item{mixprop}{Proportion that the second normal distribution holds in the mixture.}
}
\description{
PDF for Mixture of Normal Distributions
}
