% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{par2nMm}
\alias{par2nMm}
\title{transform of parameter vector to normal mixture}
\usage{
    par2nMm(par., p, k, model = c("EII", "VII", "EEI", "VEI", "EEE",
                                  "VEE", "EVI", "VVI", "EVV", "VVV"))
}
\arguments{
    \item{par.}{numeric vector of parameters}
    \item{p}{dimension of space}
    \item{model}{See description}
}
\value{
    returns this list: list(weight=w, mu=mu, Sigma=Sigma, k=k, dim=p)
}
\description{
    \code{par2nMm} returns list containing weight, mu, Sigma, k, dim
}
\details{
    this is the inverse function to nMm2par. Given a numeric vector
    dimension and cluster number this function reconstructs a normal mixture
    object.
}
