% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{par2nMmMLE}
\alias{par2nMmMLE}
\title{wrapper function for par2nMm that preprocesses for norMmixMLE}
\usage{
par2nMmMLE(par., p, k, trafo = c("clr1", "logit"), model = c("EII",
  "VII", "EEI", "VEI", "EEE", "VEE", "EVI", "VVI", "EVV", "VVV"))
}
\arguments{
\item{par.}{numeric vector of parameters}

\item{p}{dimension of space}

\item{trafo}{either "clr1" or "logit"}

\item{model}{See description}
}
\description{
\code{par2nMmMLE} returns nMm list guarantees that Sigma are sym. pos def
and alpha are nonnegative.
}
\details{
This variant of par2nMm applies exp to mixture parameters alpha and D.
to guarantee that Sigma are always sym pos def. Note here that this does not
affect weight parameter. Those may still be out of bounds of permissible 
values.
}
