% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{pmix}
\alias{pmix}
\title{CDF for a mixture of two normal distributions}
\usage{
pmix(x, mean1, sd1, mean2, sd2, mixprop)
}
\arguments{
\item{x}{Value at which to evaluate the CDF}

\item{mean1}{Mean of the first normal distribution}

\item{sd1}{Standard deviation of the first normal distribution}

\item{mean2}{Mean of the second normal distribution}

\item{sd2}{Standard deviation of the second normal distribution}

\item{mixprop}{Mixture probability}
}
\value{
Values of the CDF for the mixture of the two normal distributions.
The function is vectorized on all its parameters.
}
\description{
CDF for a mixture of two normal distributions
}
\examples{
pmix(1:3,1:3,1:3,11:13,1:3,(1:3)/4)
}
