% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormmixcode.R
\name{pnormmix}
\alias{pnormmix}
\title{Mixture of Normal Distributions}
\usage{
pnormmix(q, mean1, sd1, mean2, sd2, mixprop)
}
\arguments{
\item{q}{A given quantile for the CDF.}

\item{mean1}{Mean for the first normal distribution.}

\item{sd1}{Standard deviation for the first normal distribution.}

\item{mean2}{Mean for the second normal distribution.}

\item{sd2}{Standard deviation for the second normal distribution.}

\item{mixprop}{Proportion that the second normal distribution holds in the mixture.}

\item{x}{A given quantile for the PDF.}

\item{p}{A given probability for the inverse CDF.}

\item{n}{Number of generated numbers from the distribution.}
}
\description{
Mixture of Normal Distributions
}
