% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormmixcode.R
\name{qnormmix}
\alias{qnormmix}
\title{Inverse CDF for Mixture of Normal Distributions}
\usage{
qnormmix(p, mean1, sd1, mean2, sd2, mixprop)
}
\arguments{
\item{p}{A given probability for the inverse CDF.}

\item{mean1}{Mean for the first normal distribution.}

\item{sd1}{Standard deviation for the first normal distribution.}

\item{mean2}{Mean for the second normal distribution.}

\item{sd2}{Standard deviation for the second normal distribution.}

\item{mixprop}{Proportion that the second normal distribution holds in the mixture.}
}
\description{
Inverse CDF for Mixture of Normal Distributions
}
