% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llnorMmix.R
\name{llmvtnorm}
\alias{llmvtnorm}
\title{log-likelihood function relying on mvtnorm function}
\usage{
llmvtnorm(par., x, k, trafo = c("clr1", "logit"),
          model = c("EII", "VII", "EEI", "VEI", "EEE",
                    "VEE", "EVI", "VVI", "EVV", "VVV"))
}
\arguments{
\item{par.}{parameter vector as calculated by nMm2par}

\item{x}{matrix of samples}

\item{k}{number of cluster}

\item{trafo}{transformation of weights}

\item{model}{assumed model of the distribution}
}
\description{
\code{llmvtnorm} returns scalar value of log-likelihood
}
