% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llnorMmix.R
\name{llnorMmix}
\alias{llnorMmix}
\title{Log-likelihood of parameter vector given data}
\usage{
llnorMmix(par., x, p, k, trafo = c("clr1", "logit"), model = c("EII",
  "VII", "EEI", "VEI", "EVI", "VVI", "EEE", "VEE", "EVV", "VVV"))
}
\arguments{
\item{par.}{parameter vector}

\item{x}{sample matrix}

\item{p}{dimension}

\item{k}{number of clusters}

\item{trafo}{either centered log ratio or logit}

\item{model}{assumed distribution model of normal mixture}
}
\description{
\code{llnorMmix} returns scalar log-likelihood
}
\details{
description
}
