% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norMmix.R
\name{norMmix}
\alias{norMmix}
\title{Constructor for nMm 'objects'}
\usage{
norMmix(mu, Sigma = NULL, weight = rep(1/k, k), name = NULL,
  model = c("EII", "VII", "EEI", "VEI", "EVI", "VVI", "EEE", "VEE",
  "EVV", "VVV"))
}
\arguments{
\item{mu}{matrix of means. should mu be a vector it will assume k=1
to circumvent this behavoiur use as.matrix(mu) beforehand}

\item{Sigma}{array of covariance matrices}

\item{weight}{weights of mixture model components}

\item{name}{gives the option of naming mixture}

\item{model}{see desc}
}
\description{
\code{norMmix} returns structure containing defining parameters of a normal
mixture
}
