% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norMmix.R
\name{norMmix}
\alias{norMmix}
\title{Constructor for nMm 'objects'}
\usage{
norMmix(mu, Sigma = abind(rep(list(diag(p)), k), along = 3),
  weight = rep(1/k, k), name = NULL, model = c("EII", "VII", "EEI",
  "VEI", "EVI", "VVI", "EEE", "VEE", "EVV", "VVV"))
}
\arguments{
\item{mu}{matrix of means}

\item{Sigma}{array of covariance matrices}

\item{weight}{weights of mixture model components}

\item{name}{gives the option of naming mixture}

\item{model}{see desc}
}
\description{
\code{norMmix} returns structure containing defining parameters of a normal
mixture
}
