% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norMmixMLE.R
\name{norMmixMLE}
\alias{norMmixMLE}
\title{Maximum likelihood Estimation for normal mixture models}
\usage{
norMmixMLE(x, k, trafo = c("clr1", "logit"), model = c("EII", "VII",
  "EEI", "VEI", "EVI", "VVI", "EEE", "VEE", "EVV", "VVV"))
}
\arguments{
\item{x}{sample matrix}

\item{k}{number of clusters}

\item{trafo}{transformation to be used}

\item{model}{model to be assumed}
}
\description{
\code{norMmixMLE} returns fitted nMm obj
}
