% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{par2nMm}
\alias{par2nMm}
\title{transform of parameter vector to normal mixture}
\usage{
par2nMm(par., p, k, trafo = c("clr1", "logit"), model = c("EII", "VII",
  "EEI", "VEI", "EVI", "VVI", "EVV", "VVV"), MLE = FALSE)
}
\arguments{
\item{par.}{numeric vector of parameters}

\item{p}{dimension of space}

\item{trafo}{either "clr1" or "logit"}

\item{model}{See description}

\item{MLE}{boolean, TRUE if nonnegative entries are permitted in par.}
}
\description{
\code{par2nMm} returns list containing weight, mu, Sigma, k, dim
}
\details{
this is the inverse function to nMm2par. Given a numeric vector
dimension and cluster number this function reconstructs a normal mixture
object.
}
