% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norMmix.R
\name{plot.norMmix}
\alias{plot.norMmix}
\title{plot function for norMmix objects}
\usage{
\method{plot}{norMmix}(nMm, ...)
}
\description{
\code{plot.norMmix} returns invisibly coordinates of bounding ellipses of distribution
}
\details{
This is the S3 method for plotting norMmix objects. atm only 2 dimensional objects are supported.
}
\examples{
plot(MW212) ; points(rnorMmix(n=500, MW212))
}
