% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{pnormmix}
\alias{pnormmix}
\title{Mixture of Two Normal Distributions}
\usage{
pnormmix(x, mean1, sd1, mean2, sd2, mixprob)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{Mixture probability (weight for the first distribution, between 0 and 1).}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
Mixture of Two Normal Distributions
}
\examples{
pnormmix(1, 0, 1, 5, 1.5, 0.6)
}
