% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{qnormmix}
\alias{qnormmix}
\title{Quantile Function for a Mixture of Two Normal Distributions}
\usage{
qnormmix(p, mean1, sd1, mean2, sd2, mixprob, tol = 1e-07, maxiter = 100)
}
\arguments{
\item{p}{Numeric vector of probabilities (must be between 0 and 1).}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{Mixture probability (weight for the first distribution, between 0 and 1).}

\item{tol}{Tolerance for numerical computation (default is `1e-7`).}

\item{maxiter}{Maximum number of iterations for numerical methods (default is `100`).}
}
\description{
Quantile Function for a Mixture of Two Normal Distributions
}
\examples{
qnormmix(0.5, 0, 1, 5, 1.5, 0.6)

}
