% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{rnormmix}
\alias{rnormmix}
\title{Generate Random Samples from a Mixture of Two Normal Distributions}
\usage{
rnormmix(n, mean1, sd1, mean2, sd2, mixprob)
}
\arguments{
\item{n}{Number of random samples to generate.}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{Mixture probability (weight for the first distribution, between 0 and 1).}
}
\description{
Generate Random Samples from a Mixture of Two Normal Distributions
}
\examples{
rnormmix(1000, 0, 1, 5, 1.5, 0.6)
}
